+++ src/dawn/wire/server/ServerInstance.cpp
@@ -70,7 +70,7 @@ void Server::OnRequestAdapterCallback(RequestAdapterUserdata* data,
     // Hide features the wire cannot support.
     auto it = std::partition(features.begin(), features.end(), IsFeatureSupported);
 
-    cmd.featuresCount = static_cast<uint32_t>(std::distance(features.begin(), it));
+    cmd.featuresCount = std::distance(features.begin(), it);
     cmd.features = features.data();
 
     mProcs.adapterGetProperties(adapter, &properties);
